move delete button to separate component
